feat: some minor cleanups to the tray profiles menu layout and source code
- place boot and shutdown at the top
- add a Status "header" to the Head x/y Worker x/y section
