+++ plugins/plugin-codeflare/src/tray/menus/profiles/boot.ts
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2022 The Kubernetes Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { CreateWindowFunction } from "@kui-shell/core"
+
+import { bootIcon } from "../../icons"
+import windowOptions from "../../window"
+
+/** Handler for booting up a profile */
+async function boot(profile: string, createWindow: CreateWindowFunction) {
+  createWindow(
+    ["codeflare", "gui", "guide", "ml/ray/start/kubernetes", "--profile", profile],
+    windowOptions({ title: "Booting " + profile })
+  )
+}
+
+export default function bootMenuItem(profile: string, createWindow: CreateWindowFunction) {
+  return { label: "Boot", icon: bootIcon, click: () => boot(profile, createWindow) }
+}
